azure_iot_solution/modules/speech-to-text/requirements.txt
# azure-iothub-device-client
smbus2
requests
luma.oled
RPi.GPIO
pyaudio
pocketsphinx
